Fix wrong argument of write. Fix return type of getdents. Fix procfs_getdents. Finish implementing the base version of the ps command.

+++ libc/src/unistd/getdents.c
@@ -8,4 +8,4 @@
 #include "sys/errno.h"
 
 
-_syscall3(int, getdents, int, fd, dirent_t *, dirp, unsigned int, count)
+_syscall3(ssize_t, getdents, int, fd, dirent_t *, dirp, unsigned int, count)
